requirements.txt
Flask==1.1.1
Jinja2==2.10.3
python-Levenshtein==0.12.0
